refactor(app) : 전체 리팩터링
